Updated Main Scene
Created a Sphere game object in the Hierarchy and renamed it to Fly.
Applied the Fly script component to the Fly game object.
Created a Plane game object in the Hierarchy and renamed it to Ground.
Created 3 Cube game objects in the Hierarchy and renamed them to Obstacle (1, 2, 3).
Added Ground tag to Ground and Obstacle game objects.